pyproject.toml
[project]
name = "curation-tool"
version = "0.2.0"
description = "Image dataset curation tool with ComfyUI backend"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "pydantic>=2.0",
    "click",
    "tqdm",
    "pyyaml",
    "httpx",
    "websocket-client",
    "gradio",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[project.scripts]
curate = "curation_tool.cli:main"
